chore: polish Curve trade
