+++ src/entities/trades/curve/abi.ts
