+++ src/entities/trades/curve/constants.ts
