+++ src/entities/trades/curve/index.ts
@@ -1,1 +1,2 @@
 export * from './CurveTrade'
+export * from './types'
