+++ src/entities/trades/index.ts
@@ -4,4 +4,4 @@ export * from './interfaces/trade-options'
 export * from './interfaces/trade'
 export * from './routable-platform/routable-platform'
 export * from './routable-platform/uniswap-v2-routable-platform'
-export * from './curve-trade'
+export * from './curve'
