+++ src/entities/trades/routable-platform/routable-platform.ts
@@ -5,7 +5,7 @@ import { ChainId } from '../../../constants'
  */
 export class RoutablePlatform {
   public static readonly ZEROX = new RoutablePlatform([ChainId.MAINNET], '0x')
-  public static readonly CURVE = new RoutablePlatform([ChainId.MAINNET, ChainId.ARBITRUM_ONE, ChainId.XDAI], 'Curve')
+  public static readonly CURVE = new RoutablePlatform([ChainId.ARBITRUM_ONE, ChainId.XDAI], 'Curve')
 
   public readonly chainIds: ChainId[]
   public readonly name: string
